+++ routes/api/contacts.js
@@ -1,99 +1,93 @@
 const express = require('express');
-const Joi = require('joi');
+const ctrl = require("../../controllers/contacts")
 
-const contacts = require('../../models/contacts');
-const { HttpError } = require("../../helpers");
+// const Joi = require('joi');
+// const contacts = require('../../models/contacts');
+// const { HttpError } = require("../../helpers");
 
 const router = express.Router();
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
+// const addSchema = Joi.object({
+//   name: Joi.string().required(),
+//   email: Joi.string().required(),
+//   phone: Joi.string().required(),
+// });
 
-router.get('/', async (req, res, next) => {
-  try {
-    const result = await contacts.listContacts();
-    res.json(result);
-  } catch (error) {
-    next(error);
-    // res.status(500).json({
-    //   message: "Server error",
-    // })
-  }
-  
-})
+router.get('/', ctrl.getAll);
 
-router.get('/:contactId', async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.getContactById(contactId);
-    if (!result) {
-      throw HttpError(404, "Not found");
-      // const error = new Error("Not found");
-      // error.status = 404;
-      // throw error;
+// router.get('/', async (req, res, next) => {
+//   try {
+//     const result = await contacts.listContacts();
+//     res.json(result);
+//   } catch (error) {
+//     next(error);
+//   }
+// });
 
-      // return res.status(404).json({
-      //   message: "Not found",
-      // }); 
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-    // const { status = 500, message = "Server error" } = error;
-    // res.status(status).json({
-    //   message ,
-    // })
-  }
-})
+router.get('/:contactId', ctrl.getById);
 
-router.post('/', async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, "Missing required name field");
-    }
-    const result = await contacts.addContact(req.body);
-    console.log(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-})
+// router.get('/:contactId', async (req, res, next) => {
+//   try {
+//     const { contactId } = req.params;
+//     const result = await contacts.getContactById(contactId);
+//     if (!result) {
+//       throw HttpError(404, "Not found");
+//     }
+//     res.json(result);
+//   } catch (error) {
+//     next(error);
+//   }
+// })
 
-router.delete('/:contactId', async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.removeContact(contactId);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json({
-      message: "Contact deleted",
-    });
-  } catch (error) {
-    next(error);
-  }
-})
+router.post('/', ctrl.add);
 
-router.put('/:contactId', async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, "Missing fields");
-    }
-    const { contactId } = req.params;
-    const result = await contacts.updateContact(contactId, req.body);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    console.log(result);
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-})
+// router.post('/', async (req, res, next) => {
+//   try {
+//     const { error } = addSchema.validate(req.body);
+//     if (error) {
+//       throw HttpError(400, "Missing required name field");
+//     }
+//     const result = await contacts.addContact(req.body);
+//     res.status(201).json(result);
+//   } catch (error) {
+//     next(error);
+//   }
+// })
+
+router.delete('/:contactId', ctrl.deleteById);
+
+// router.delete('/:contactId', async (req, res, next) => {
+//   try {
+//     const { contactId } = req.params;
+//     const result = await contacts.removeContact(contactId);
+//     if (!result) {
+//       throw HttpError(404, "Not found");
+//     }
+//     res.json({
+//       message: "Contact deleted",
+//     });
+//   } catch (error) {
+//     next(error);
+//   }
+// })
+
+router.put('/:contactId', ctrl.updateById);
+
+// router.put('/:contactId', async (req, res, next) => {
+//   try {
+//     const { error } = addSchema.validate(req.body);
+//     if (error) {
+//       throw HttpError(400, "Missing fields");
+//     }
+//     const { contactId } = req.params;
+//     const result = await contacts.updateContact(contactId, req.body);
+//     if (!result) {
+//       throw HttpError(404, "Not found");
+//     }
+//     res.json(result);
+//   } catch (error) {
+//     next(error);
+//   }
+// })
 
 module.exports = router;
